grp1/grp1/requirements.txt
opencv-python
numpy
gpiozero
Adafruit-PCA9685
RPi.GPIO
